Improve command line parsing
Instead of allowing only one quote at the beginning and one at the end
of the command name, now we're reading the command line character by
character, also advancing in the CRT-parsed command name and skipping
any number of quotes.

As far as I can tell, there is no other hole in it, but just to be sure,
any parsing failure is now fatal.